backend/requirements.txt
asgiref==3.11.1
asttokens==3.0.1
attrs==26.1.0
decorator==5.2.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
executing==2.2.1
gunicorn==22.0.0
inflection==0.5.1
ipython==8.26.0
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
matplotlib-inline==0.2.1
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.12.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
uritemplate==4.2.0
wcwidth==0.6.0
openpyxl
drf-nested-routers
